microservices/notification-service/requirements.txt
Flask==3.0.0
kafka-python==2.0.2
prometheus-flask-exporter==0.23.0
gunicorn==21.2.0
